group_path_count, end_path_count use size_t

+++ include/sta/SearchClass.hh
@@ -114,7 +114,7 @@ typedef Vector<MaxSkewCheck*> MaxSkewCheckSeq;
 typedef StringSet PathGroupNameSet;
 typedef Vector<PathEnd*> PathEndSeq;
 typedef Vector<Arrival> ArrivalSeq;
-typedef Map<Vertex*, int> VertexPathCountMap;
+typedef Map<Vertex*, size_t> VertexPathCountMap;
 typedef UnorderedMap<Tag*, size_t, TagMatchHash, TagMatchEqual> PathIndexMap;
 typedef Vector<Slack> SlackSeq;
 typedef Delay Crpr;
